ngapus edit, dicampur

+++ laporanKas.php
@@ -9,10 +9,21 @@ include "koneksi.php";
 	$pengeluaran = $data_total['totalKeluar'] ?? 0;
 	$saldo  = $pemasukan - $pengeluaran;
 
+    if (isset($_GET['aksi']) && $_GET['aksi'] == 'tambahKas') {
+        mysqli_query($connect, "INSERT INTO kas(tanggal, jenis, kategori, nominal, keterangan) VALUES ('{$_POST['tanggal']}','{$_POST['jenis']}','{$_POST['kategori']}', '{$_POST['nominal']}','{$_POST['keterangan']}')");
+        header("Location: laporanKas.php");
+        exit;
+    }
+    if (isset($_GET['aksi']) && $_GET['aksi'] == 'simpanEdit') {
+        $id = $_GET['id'];
+        mysqli_query($connect, "UPDATE kas SET tanggal='{$_POST['tanggal']}', jenis='{$_POST['jenis']}', kategori='{$_POST['kategori']}', nominal='{$_POST['nominal']}', keterangan='{$_POST['keterangan']}' WHERE kasID = '$id'");
+        header("Location: laporanKas.php");
+        exit;
+    }
 	if (isset($_GET['aksi']) && $_GET['aksi'] == 'hapus') {
 	    $id = $_GET['id'];
 	    mysqli_query($connect, "DELETE FROM kas where kasID = '$id'");
-	    header("Location: kas.php");
+	    header("Location: laporanKas.php");
 	    exit;
 	}
 	if(isset($_GET['bulan'])&&isset($_GET['tahun'])){
@@ -40,7 +51,51 @@ include "koneksi.php";
 	        from kas");
 	}
 ?>
+    <?php if(isset($_GET['aksi'])&& $_GET['aksi'] == 'editKas'): 
+        $id = $_GET['id'];
+        $query = mysqli_query($connect, "SELECT * from kas where kasID = '$id' ");
+        $data = mysqli_fetch_array($query); ?>
+        	<form action="laporanKas.php?aksi=simpanEdit&id=<?=$data['kasID']?>" method="POST">
+                <label for="id">ID : </label>
+                <i><?= $data['kasID']?></i><br>
+                <label for="jenis">Jenis</label>
+                <select name="jenis" required>
+                    <option value="" disabled selected>Jenis Kas...</option>
+                    <option value="Kas Masuk" <?= ($data['jenis'] == 'Kas Masuk')?'selected':''; ?>>Kas Masuk</option>
+                    <option value="Kas Keluar" <?= ($data['jenis'] == 'Kas Keluar')?'selected':''; ?>>Kas Keluar</option>
+                </select> <br>
+                <label for="kategori">Kategori</label>
+                <input type="text" name="kategori" value="<?= $data['kategori'] ?>" required> <br>
+                <label for="tanggal">Tanggal Transfer</label>
+                <input type="date" name="tanggal" value="<?= $data['tanggal'] ?>" required> <br>
+                <label for="nominal">Nominal (Rp)</label>
+                <input type="number" name="nominal" value="<?= $data['nominal'] ?>" required> <br>
+                <label for="keterangan">Keterangan</label>
+                <input type="text" name="keterangan" value="<?= $data['keterangan'] ?>" required> <br>
+                <button type="submit">Simpan Perubahan</button>
+            </form>
 
+
+    <?php elseif(isset($_GET['aksi'])&& $_GET['aksi'] == 'inputKas'):  ?>
+            <form action="laporanKas.php?aksi=tambahKas" method="POST">
+                <label for="jenis">Jenis</label>
+                <select name="jenis" required>
+                    <option value="" disabled selected>Jenis Kas...</option>
+                    <option value="Kas Masuk">Kas Masuk</option>
+                    <option value="Kas Keluar">Kas Keluar</option>
+                </select> <br>
+                <label for="kategori">Kategori</label>
+                <input type="text" name="kategori" required> <br>
+                <label for="tanggal">Tanggal Transfer</label>
+                <input type="date" name="tanggal" required> <br>
+                <label for="nominal">Nominal (Rp)</label>
+                <input type="number" name="nominal" required> <br>
+                <label for="keterangan">Keterangan</label>
+                <input type="text" name="keterangan"> <br>
+                <button type="submit">Input kas</button>
+            </form>
+
+	<?php else:?>
 	<h3>Pemasukan: <?=$pemasukan?></h3>
 	<h3>Pengeluaran: <?=$pengeluaran?></h3>
 	<h3>Saldo: <?=$saldo?></h3>
@@ -64,7 +119,7 @@ include "koneksi.php";
 	    <button type="submit">Tampilkan</button>
 	    <a href="?" style="padding: 3px 10px; background: black; color: white; text-decoration: none; border-radius: 3px;">Reset</a>
 	</form>
-  	<a href="edit.php?aksi=inputKas" style="padding: 5px 10px; background: blue; color: white; text-decoration: none; border-radius: 3px;">
+  	<a href="laporanKas.php?aksi=inputKas" style="padding: 5px 10px; background: blue; color: white; text-decoration: none; border-radius: 3px;">
   		input
 	</a><br><br> 
 	<table border="1" cellpadding="8">
@@ -88,7 +143,7 @@ include "koneksi.php";
 				<td style="color: <?= ($data['jenis'] == 'Kas Masuk') ? 'green' : 'red'; ?>"><?= "Rp " . number_format($data['nominal'], 2, ',', '.'); ?></td>
 				<td><?= "Rp " . number_format($data['saldo'], 2, ',', '.'); ?></td>
 				<td>
-            	<a href="edit.php?aksi=editKas&id=<?= $data['kasID']; ?>" 
+            	<a href="laporanKas.php?aksi=editKas&id=<?= $data['kasID']; ?>" 
                  style="padding: 5px 10px; background: black; color: white; text-decoration: none; border-radius: 3px;">
                  Edit
             	</a>  
@@ -101,3 +156,5 @@ include "koneksi.php";
 			</tr>
 		<?php } ?>
 	</table>
+
+	<?php endif; ?>
